iproto: implement graceful handling of receiving eof when reading
We can receive eof in two cases: when connection closed by client or when
client send SHUT_WR. In second case we should processed all requests and
close connection only after it. So now we don't close the connection after
receiving the eof, but only stop read watcher. We close connection after
processing all requests or in case when we getting error when writing.
Part of #5924