test: adapt disabled tests from PUC-Rio
Version and status are printed in stdout instead stderr
since LuaJIT-2.0.0-beta11 (as it is not an error message). This behavior
is the same as in Lua 5.2, so necessary changes in tests can be
adapted from PUC-Rio Lua 5.2 test suite.

However, those tests are disabled for Tarantool since its interactive
shell doesn't support '=' at the beginning of statements.
Here is an example:
========================================================================
$ luajit
> = (6*2-6) -- ===
6

$ tarantool
tarantool> = (6*2-6) -- ===
---
- error: ' (6*2-6) -- ===:1: unexpected symbol near ''='''
...
========================================================================

Part of tarantool/tarantool#5870