uri: implement ability to parse URI query paramters
Previously, the part of the URI string that comes after the first '?' was
saved as a single unit in the field `query` as a part of `struct uri`. Now
an array of URI query paramters has been added to the uri structure, and
query is subjected to the additional parsing and URI query parameters are
extracted from it. The separator symbol between URI query parameters is '&',
the separator symbol between parameter and it's value is '='. For example
URI = "/tmp/unix.sock?q1=v11&q1=v12&q2=v2" has two parameters: parameter
'q1' with two values 'v11' and 'v12' and parameter 'q2' with value 'v2'.

Part of #5928