box: add OPT_CUSTOM to opt_def
opt_def is an option parser from msgpack by given scheme.
The scheme consists of set of predefined types that the parser
can handle (something like int, enum, str). The problen is that
those predefined types must be generic, but there are cases when
a specific unusual must be parsed.

This patch introduces OPT_CUSTOM type that uses arbitrary function
callback and can be used for any non-generic option value.