box: implement field constraints
Introduce field constraints - limitaions for particular fields.
Each constraint must refer to a function in _func space. For the
first step we expect lua functions with body there.

Field constraint checks are very close to field type checks, so
it's natural to implement them in tuple formats. On the other hand
tuple formats belong to tuple library, that does not include
functions (func.h/c etc), so constraints are split into two parts:
- a part in tuple library that implements arbitrary constraints
 with pointers to functions that actualy check constraints.
- a part in box library which uses the part above, sets particular
 check functions and handles alter etc.

There are two not-so-obvious problems that are solved here:
 - Functions in _func space must be preserved while used by such
 constraints. Func pinning is used for this purpose.
 - During initial recovery constraits are created before _func
 space recovery, so we have to pospone constraint initialization.
 Func subscription is used for that with the assumption a snapshot
 is guaranteed to be correct and needed functions will arise later.