alter: introduce local space:upgrade()
Upgrade operation allows users to change format (which can be
incompatible with old one) and convert data according to it. To achieve
this we introduce new system space _space_upgrade. Each in-progress
upgrade inserts entry to it. Tuple of _space_upgrade contains:
space_id -- id of the space being updated;
status   -- string containing status of update (in-progress, test or fail);
func_id  -- id of the upgrade function;
format   -- map corresponding to the new space format.

Upgrade function should accept one argument - tuple to be converted
(in fact there can be more arguments, but they are always ignored and
they are nils) and return new (converted) tuple (should be an array).
Function also has to be written in Lua, persistent, sandboxed and
deterministic. New format may not be compatible with the old one. The
only restriction is it must not interfere with indexed field.

Upgrade features three modes: "test", "notest" and casual "upgrade".
In the test mode no data changes are made. We simply iterate over the
space and check if upgrade function is applicable and resulting tuple
satisfies new format. Still the entry is inserted into _space_upgrade -
it just simplifies metadata parsing (space format etc).
In "notest" mode we in the same way process insertion to _space_upgrade
and then alter space format (by casual space:alter()). If it does not
contradicts with indexed field types, then we start upgrade. Once in
several upgraded tuples (10 in debug, 1000 in release) we commit
new tuples and yield. If upgrade fails due to whatever reason, operation
is aborted and its status is changed to "error" (by updating
corresponding entry in the system space). In such state space is still
able to handle dml operations which don't break new format. To take out
of error state user has to run space:upgrade() again with another
function or format.
"upgrade" mode is a mix of "test" and "notest" modes. In case test
finishes with success - causal upgrade starts.

During real upgrade space is locked in terms of any other DDL operation
(including other upgrades).

Part of #xxxx