test: fix flaky read_only_reason test
This is a second attempt to stabilize #5568 test since
5105c2d794b43d6733b3cbcbbbe418fe02d56d79 ("test: fix flaky
read_only_reason test").

This time the error sometimes appeared in the test case
test_read_only_reason_synchro.

Here a replica during one exec() saw it is read-only because the
synchro queue is foreign. But on the next exec() it suddenly
wasn't read-only anymore (box.info.ro_reason was nil).

So it was like the replica demoted the old leader.

The actual reason was that the test case runs after
test_read_only_reason_election_has_leader. Here the master
instance in the end makes box.ctl.demote().

But the quorum is 1 by default. So when the previous test case
ended, demote() not necessarily reached the replica.

The failure of test_read_only_reason_synchro happened when that
previous test case's demote() reached the replica between these 2
execs.

Note, that even though test_read_only_reason_synchro does another
promote() and hence bumps the term and thus should have ignored
the old demote, ironically it had the same problem. New promote()
was done on the same master as the previous demote() and also had
quorum 1. So demote() reached the replica before promote() anyway.
Even if box.ctl.promote() was already returned on master.

The patch makes the tests use proper quorum values. So when
promote follows demote, it is guaranteed that demote is finished
first, including being received and applied by the replica.

Also the patch adds explicit waiting for the replica to become RW
in the end of the test cases. Also to ensure the demote ended.

Closes #6670